fix(hooks): preserve sub-second shell timeout_secs precision (#325)
`time.Duration(t) * time.Second` for a float64 t casts t to int first,
truncating any sub-second value (e.g. 0.5) to zero before the
multiplication. The resulting zero duration trips the context deadline
immediately, so a hook authored with `timeout_secs: 0.5` would
silently fire and return timed_out=true in microseconds — with no
indication of why the script never ran.

Switch the float64 branch to compute nanoseconds directly via
`time.Duration(t * float64(time.Second))`. The int branch is unaffected
because integer seconds round-trip cleanly. Add a regression test for
0.5s timeout vs `sleep 2`, plus a subtest suite for the non-positive
fallthrough case documenting the intentional 30s-default behavior.

+++ internal/hooks/shell.go
@@ -64,13 +64,20 @@ func shellHandler(ctx context.Context, h Hook, ec EventContext) (Result, error) 
 	// context deadline. Cap any caller-supplied seconds at a defensible
 	// upper bound (24 hours) so misconfigured hooks fail at parse time
 	// instead of producing the surprising "immediate timeout" symptom.
+	//
+	// Sub-second floats need nanosecond math: `time.Duration(0.5) *
+	// time.Second` casts 0.5 to int(0) first, which is also an
+	// immediate-timeout footgun. Multiply in float space and cast once.
+	// Non-positive values fall through to the 30s default — surfacing an
+	// explicit error would break any hook author who left timeout_secs
+	// unset or zeroed.
 	const maxTimeoutSecs = 24 * 60 * 60
 	timeout := 30 * time.Second
 	if t, ok := h.HandlerConfig["timeout_secs"].(float64); ok && t > 0 {
 		if t > maxTimeoutSecs {
 			t = maxTimeoutSecs
 		}
-		timeout = time.Duration(t) * time.Second
+		timeout = time.Duration(t * float64(time.Second))
 	}
 	if t, ok := h.HandlerConfig["timeout_secs"].(int); ok && t > 0 {
 		if t > maxTimeoutSecs {
